Implemented data extensions source plugin
