+++ src/main/java/io/cdap/plugin/sfmc/DataExtensionOutputFormat.java
@@ -18,6 +18,7 @@ package io.cdap.plugin.sfmc;
 
 import com.exacttarget.fuelsdk.ETSdkException;
 import io.cdap.cdap.api.data.format.StructuredRecord;
+import io.cdap.plugin.sfmc.common.DataExtensionClient;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.io.NullWritable;
 import org.apache.hadoop.mapreduce.JobContext;
