+++ widgets/SalesforceDataExtension-batchsource.json
@@ -2,7 +2,7 @@
   "metadata": {
     "spec-version": "1.0"
   },
-  "display-name": "Salesforce",
+  "display-name": "Salesforce Marketing",
   "configuration-groups": [
     {
       "label": "Basic",
@@ -17,59 +17,62 @@
         },
         {
           "widget-type": "select",
-          "label": "Data retrieval mode",
+          "label": "Data Retrieval Mode",
           "name": "queryMode",
           "widget-attributes": {
             "values": [
-              "Multi-Object",
-              "Single-Object"
+              "Multi Object",
+              "Single Object"
             ],
-            "default": "Single-object"
+            "default": "Single Object"
           }
         }
       ]
     },
     {
-      "label": "Single-Object Mode",
-      "properties": [
-        {
-          "widget-type": "select",
-          "label": "Object Name",
-          "name": "objectName",
-          "widget-attributes": {
-            "values": [
-              "DataExtension",
-              "Campaign",
-              "Account",
-              "Email"
-            ],
-            "default": "DataExtension"
-          }
-        }
-      ]
-    },
-    {
-      "label": "Multi-Object Mode",
+      "label": "Single Object Retrieval",
       "properties": [
         {
           "widget-type": "textbox",
-          "label": "List Of Object Name",
-          "name": "multiObjectName",
+          "label": "Data Extension External Key",
+          "name": "dataExtensionKey",
           "widget-attributes": {
-            "placeholder": "Salesforce object name from which data to be fetched"
+            "placeholder": "External key of the Data Extension from which data to be read."
           }
         }
       ]
     },
     {
-      "label": "Credentials",
+      "label": "Multi Object Retrieval",
+      "properties": [
+        {
+          "widget-type": "csv",
+          "label": "Data Extension External Keys",
+          "name": "dataExtensionKeyList",
+          "widget-attributes": {
+            "placeholder": "External keys of the Data Extensions from which data to be read."
+          }
+        },
+        {
+          "widget-type": "textbox",
+          "label": "Table Name Field",
+          "name": "tableNameField",
+          "widget-attributes": {
+            "placeholder": "The name of the field that holds the table name.",
+            "default": "tablename"
+          }
+        }
+      ]
+    },
+    {
+      "label": "Authentication",
       "properties": [
         {
           "widget-type": "textbox",
           "label": "Client ID",
           "name": "clientId",
           "widget-attributes" : {
-            "placeholder": "Salesforce Client ID"
+            "placeholder": "Salesforce Marketing Cloud Client ID"
           }
         },
         {
@@ -77,15 +80,15 @@
           "label": "Client Secret",
           "name": "clientSecret",
           "widget-attributes" : {
-            "placeholder": "Salesforce Client Secret Key"
+            "placeholder": "Salesforce Marketing Cloud Client Secret"
           }
         },
         {
           "widget-type": "textbox",
-          "label": "AUTH API Endpoint",
+          "label": "Authentication Base URI",
           "name": "authEndpoint",
           "widget-attributes" : {
-            "placeholder": "Salesforce REST API Endpoint e.g. https://instance.auth.marketingcloudapis.com"
+            "placeholder": "Salesforce Auth API Endpoint e.g. https://instance.auth.marketingcloudapis.com"
           }
         },
         {
@@ -98,7 +101,7 @@
         },
         {
           "widget-type": "textbox",
-          "label": "REST API Endpoint",
+          "label": "REST API Base URI",
           "name": "restEndpoint",
           "widget-attributes" : {
             "placeholder": "Salesforce REST API Endpoint e.g. https://instance.rest.marketingcloudapis.com"
@@ -110,31 +113,14 @@
       "label": "Advanced",
       "properties": [
         {
-          "widget-type": "select",
-          "label": "Type of values",
-          "name": "valueType",
-          "widget-attributes": {
-            "values": [
-              "Actual",
-              "Display"
-            ],
-            "default": "Actual"
-          }
-        },
-        {
-          "widget-type": "textbox",
-          "label": "Start Date",
-          "name": "startDate",
+          "widget-type": "number",
+          "label": "Page Size",
+          "name": "pageSize",
           "widget-attributes" : {
-            "placeholder": "Start date to be used to filter the data"
-          }
-        },
-        {
-          "widget-type": "textbox",
-          "label": "End Date",
-          "name": "endDate",
-          "widget-attributes" : {
-            "placeholder": "End date to be used to filter the data"
+            "default": "500",
+            "min": "1",
+            "max": "2500",
+            "placeholder": "The number of records to be fetched in one page."
           }
         }
       ]
